float

+++ ex3.py
@@ -2,16 +2,16 @@
 print("I will now count my chickens")
 
 #print out string Hens and the number of hens 
-print("Hens", 25+30/6)
+print("Hens", float(25+30/6))
 
 #print out roosters, and the number of roosters 
-print("Roosters", 100 - 25 *3 % 4)
+print("Roosters", float(100 - 25 *3 % 4))
 
 #print out string Now I will count the eggs
 print("Now I will count the eggs:")
 
 #print out the numbers of egg
-print(3 + 2 + 1 - 5 + 4 % 2 - 1 / 4 + 6)
+print(float(3 + 2 + 1 - 5 + 4 % 2 - 1 / 4 + 6))
 
 #print out string is it true that 3 + 2 < 5 - 7
 print("Is it true that 3 + 2 < 5 - 7?")
@@ -20,9 +20,9 @@ print("Is it true that 3 + 2 < 5 - 7?")
 print(3 + 2 < 5 - 7)
 
 #print out string what is 3 + 2? and the answer
-print("What is 3 + 2?",  3 + 2)
+print("What is 3 + 2?",  float(3 + 2))
 #print out string what is 5 - 7? and the result
-print("What is 5 - 7?", 5 -7)
+print("What is 5 - 7?", float(5 -7))
 
 #print out string oh, that's why it's false
 print("Oh, that's why it's False.")
